Allow some 'views' to not have a lifetime
This is in the case where a 'view' (it is really some sort of new
category) contains only owned types that are not zerocopy, and so
doesn't need a lifetime.

I frankly do not know if this is a good idea or not; So far it is
basically just a thing that i have done, because I had backed myself
into an inconsistency.