[subset] Checkpoint initial gpos subsetting work
tldr; this compiles, but I don't have compile types in place for "making
an actual font" yet, and I also have *zero* clue whether or not GPOS is
currently compiling correctly, I just sort-of know that it is compiling,
period. Now I need to go back and fill in those two gaps.